djangostorioapp/requirements.txt
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.0.1
django-storages==1.11.1
djangorestframework==3.12.2
google-api-core==1.26.3
google-auth==1.29.0
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
gunicorn==20.0.4
idna==2.10
packaging==20.9
Pillow==8.1.2
protobuf==3.15.8
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
django-allauth==0.42.0
